+++ services/kuber/manifests/claudie-defaults.yaml
