+++ services/kuber/manifests/manifests.go
